+++ service/src/main/java/com/g2/scheduleservice/application/impl/ScheduleServiceImpl.java
@@ -12,9 +12,6 @@ import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.time.temporal.ChronoField;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 @Service
@@ -24,11 +21,11 @@ public class ScheduleServiceImpl implements ScheduleService {
     private final TimeEditClient timeEditClient;
 
     @Override
-    public CourseOccasionScheduleResponse getReservations(int objectId, LocalDate firstDate, LocalDate lastDate) {
-        System.out.println("FirstDate " + firstDate + " becomes "+ toDateint(firstDate));
+    public CourseOccasionScheduleResponse getReservations(long objectId, LocalDate firstDate, LocalDate lastDate) {
+        System.out.println("FirstDate " + firstDate + " becomes " + toDateint(firstDate));
         val timeEditResponse = timeEditClient.getObject(objectId, toDateint(firstDate), toDateint(lastDate));
 
-        return  toCourseOccasionScheduleResponse(timeEditResponse.getBody());
+        return toCourseOccasionScheduleResponse(timeEditResponse.getBody());
 
     }
 
@@ -37,17 +34,17 @@ public class ScheduleServiceImpl implements ScheduleService {
         System.out.println("SNSUSJK");
         return CourseOccasionScheduleResponse.builder()
                 .reservations(reservations)
-                .CourseCode(reservations.get(0).getDescription().substring(0,5))
+                .CourseCode(reservations.get(0).getDescription().substring(0, 5))
                 .build();
     }
 
-    private ReservationResponse toReservationResponse(TimeEditReservationResponse input){
-        System.out.println("JOIJH"+input);
+    private ReservationResponse toReservationResponse(TimeEditReservationResponse input) {
+        System.out.println("JOIJH" + input);
         return ReservationResponse.builder()
                 .contactName(input.getColumns().get(2))
                 .title(input.getColumns().get(3))
                 .distanceUrl(input.getColumns().get(7))
-                .eventUrl("https://cloud.timeedit.net/ltu/web/schedule1/ri.json?h=t&sid=3&objects="+input.getId()+".28&ox=0&types=0&fe=0")
+                .eventUrl("https://cloud.timeedit.net/ltu/web/schedule1/ri.json?h=t&sid=3&objects=" + input.getId() + ".28&ox=0&types=0&fe=0")
                 .location(input.getColumns().get(1))
                 .startTime(LocalDateTime.of(input.getStartdate(), input.getStarttime()))
                 .endTime(LocalDateTime.of(input.getEnddate(), input.getEndtime()))
@@ -56,8 +53,8 @@ public class ScheduleServiceImpl implements ScheduleService {
 
     }
 
-    private int toDateint(LocalDate input){
-        int year = input.getYear()-2000;
+    private int toDateint(LocalDate input) {
+        int year = input.getYear() - 2000;
         int month = input.getMonthValue();
         int day = input.getDayOfMonth();
